refactor: 도메인에 대한 테스트 코드 분리
