+++ src/test/kotlin/lotto/LottoMachineTest.kt
@@ -1,0 +1,39 @@
+package lotto
+
+import io.kotest.core.spec.style.StringSpec
+import io.kotest.matchers.shouldBe
+import lotto.domain.Lotto
+import lotto.domain.LottoMachine
+import lotto.domain.LottoNumber
+import org.assertj.core.api.Assertions
+
+class LottoMachineTest : StringSpec({
+
+    "로또 1장의 가격은 1000원 이다" {
+
+        Assertions.assertThatExceptionOfType(IllegalArgumentException::class.java)
+            .isThrownBy { LottoMachine(100) }
+    }
+
+    "로또 하나의 숫자는 6개 이다" {
+        val numbers = (1..7).map { LottoNumber(it) }
+
+        Assertions.assertThatExceptionOfType(IllegalArgumentException::class.java)
+            .isThrownBy { Lotto(numbers.toSet()) }
+    }
+
+    "생성한 난수 6개는 중복이 없어야 한다." {
+        val lottoMachine = LottoMachine(1000)
+        val lottoList = lottoMachine.publishLotto()
+
+        lottoList.size shouldBe lottoList.distinct().size
+    }
+
+    "로또 n장에 따라 난수 6개씩 n개 생성" {
+        val lottoMachine = LottoMachine(14000)
+
+        val lottoList = lottoMachine.publishLotto()
+
+        lottoList.size shouldBe lottoMachine.purchaseCount
+    }
+})
